+++ NEWS.md
@@ -37,6 +37,7 @@
 - Fixed viewing or blaming file on GitHub for a newly created branch. (#9798)
 - Fixed an issue on macOS where '-ne' was erroneously printed to the console with certain versions of Bash. (#13809)
 - Fixed an issue where attempts to open files containing non-ASCII characters from the Files pane could fail on Windows. (#13855, #12467)
+- Fixed an issue where color highlight for Console input could not be disabled. (#13118)
 
 #### Posit Workbench
 - Fixed opening job details in new windows more than once for Workbench jobs on the homepage. (rstudio/rstudio-pro#5179)